+++ packages/agent-runtime/src/types/index.ts
@@ -1,6 +1,7 @@
 export * from './event';
 export * from './generalAgent';
 export * from './graph';
+export * from './hooks';
 export * from './instruction';
 export * from './runtime';
 export * from './state';
